+++ src/app/layout/sidebar/Menu.jsx
@@ -27,7 +27,7 @@ class Menu extends React.Component {
           <MenuItem name={ window.L.t('Saved Searches') } target="/saved-searches" icon={<MdGroup />} />
           <MenuItem name={ window.L.t('Search Creator') } target="/search-creator" icon={<MdBuild />} />
           {/*<MenuItem name={ window.L.t('Tag Manager') } target="/tag-manager" icon={<MdSettings />}/>*/}
-          <MenuItem name="Ask" target="/asks" icon={<MdArtTrack />}/>
+          <MenuItem name="View Forms" target="/forms" icon={<MdArtTrack />}/>
           <MenuItem name={ window.L.t('About') } target="/about" icon={<MdInfoOutline />}/>
           {/*<MenuItem name="Settings" target="/settings" icon={<MdSettings />} />*/}
         </ul>
